chore: extract reusable virtua mock for tests
